+++ database/migrations/2024_07_12_115053_add_payment_balance_to_reports_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('reports', function (Blueprint $table) {
-            $table->integer('payment_balance')->nullable();
+            $table->float('payment_balance', 8, 2)->nullable();
         });
     }
